add JSON module and activation configuration after module check

+++ cui-php-apache2/cui-php-apache2.sh
@@ -232,16 +232,13 @@ fi
 
 # -----------------------------------------------------------------------------
 # MYSQL
+rm  -f /etc/php/conf.d/mysql.ini
+rm  -f /etc/php/conf.d/mysqli.ini
+rm  -f /etc/php/conf.d/pdo_mysql.ini
 if [ "$PHP_EXT_MYSQL" = "yes" ] ; then
-	if ! apk info -q -e php-mysql; then
-		apk add -q php-mysql
-	fi
-	if ! apk info -q -e php-mysqli; then
-		apk add -q php-mysqli
-	fi
-	if ! apk info -q -e php-pdo_mysql; then
-		apk add -q php-pdo_mysql
-	fi
+    apk info -q -e php-mysql || apk add -q php-mysql
+    apk info -q -e php-mysqli || apk add -q php-mysqli
+    apk info -q -e php-pdo_mysql || apk add -q php-pdo_mysql
     if [ -z "$PHP_EXT_MYSQL_SOCKET" -a -z "$PHP_EXT_MYSQL_HOST" ] ; then
         [ -e "/run/mysqld/mysqld.sock" ] && PHP_EXT_MYSQL_SOCKET="/run/mysqld/mysqld.sock"
     fi
@@ -250,7 +247,9 @@ if [ "$PHP_EXT_MYSQL" = "yes" ] ; then
     else
         [ -z "$PHP_EXT_MYSQL_PORT" ] && PHP_EXT_MYSQL_PORT="3306"
     fi
-    cat >/etc/php/conf.d/mysql.ini <<EOF
+
+	if apk info -q -e php-mysql; then
+        cat >/etc/php/conf.d/mysql.ini <<EOF
 extension=mysql.so
 [mysql]
 mysql.allow_local_infile=On
@@ -266,15 +265,17 @@ mysql.default_password=
 mysql.connect_timeout=60
 mysql.trace_mode=Off
 EOF
-
-    cat >/etc/php/conf.d/pdo_mysql.ini <<EOF
+    fi
+	if apk info -q -e php-mysqli; then
+        cat >/etc/php/conf.d/pdo_mysql.ini <<EOF
 extension=pdo_mysql.so
 [pdo_mysql]
 pdo_mysql.cache_size=2000
 pdo_mysql.default_socket=${PHP_EXT_MYSQL_SOCKET}
 EOF
-
-    cat >/etc/php/conf.d/mysqli.ini <<EOF
+    fi
+	if apk info -q -e php-pdo_mysql; then
+        cat >/etc/php/conf.d/mysqli.ini <<EOF
 extension=mysqli.so
 [mysqli]
 mysqli.max_persistent=-1
@@ -295,16 +296,16 @@ mysqlnd.collect_memory_statistics=Off
 ;mysqlnd.net_cmd_buffer_size=2048
 ;mysqlnd.net_read_buffer_size=32768
 EOF
-else
-    rm  -f /etc/php/conf.d/mysql.ini
-    rm  -f /etc/php/conf.d/mysqli.ini
-    rm  -f /etc/php/conf.d/pdo_mysql.ini
+    fi
 fi
 
-# -----------------------------------------------------------------------------
+# ------------------------------------------------------------------------------
 # INTERBASE
+rm -f /etc/php/conf.d/interbase.ini
 if [ "$PHP_EXT_INTER" = "yes" ] ; then
-    cat >/etc/php/conf.d/interbase.ini <<EOF
+#    apk info -q -e php-interbase || apk add -q php-interbase
+	if apk info -q -e php-interbase; then
+        cat >/etc/php/conf.d/interbase.ini <<EOF
 ;extension=interbase.so
 ;extension=pdo_firebird.so
 [interbase]
@@ -329,17 +330,16 @@ ibase.dateformat = "%Y-%m-%d"
 ; Default time format.
 ibase.timeformat = "%H:%M:%S"
 EOF
-else
-    rm -f /etc/php/conf.d/interbase.ini
+    fi
 fi
 
 # -----------------------------------------------------------------------------
 # MSSQL
+rm -f /etc/php/conf.d/mssql.ini
 if [ "$PHP_EXT_MSSQL" = "yes" ] ; then
-	if ! apk info -q -e php-mssql; then
-		apk add -q php-mssql
-	fi	
-    cat >/etc/php/conf.d/mssql.ini <<EOF
+    apk info -q -e php-mssql || apk add -q php-mssql
+	if apk info -q -e php-mssql; then
+        cat >/etc/php/conf.d/mssql.ini <<EOF
 extension=mssql.so
 [mssql]
 ; Allow or prevent persistent links.
@@ -379,20 +379,18 @@ mssql.secure_connection = Off
 ; This is only used when compiled with FreeTDS
 ;mssql.charset = "ISO-8859-1"
 EOF
-else
-    rm -f /etc/php/conf.d/mssql.ini
+    fi
 fi
 
 # -----------------------------------------------------------------------------
 # POSTGRESQL
+rm -f /etc/php/conf.d/pqsql.ini
+rm -f /etc/php/conf.d/pdo_pgsql.ini
 if [ "${PHP_EXT_PGSQL}" = "yes" ] ; then
-	if ! apk info -q -e php-pgsql; then
-		apk add -q php-pgsql
-	fi
-	if ! apk info -q -e php-pdo_pgsql; then
-		apk add -q php-pdo_pgsql
-	fi
-    cat >/etc/php/conf.d/pqsql.ini <<EOF
+    apk info -q -e php-pgsql || apk add -q php-pgsql
+    apk info -q -e php-pdo_pgsql || apk add -q php-pdo_pgsql
+	if apk info -q -e php-pgsql; then
+        cat >/etc/php/conf.d/pqsql.ini <<EOF
 extension=pgsql.so
 [PostgresSQL]
 ; Allow or prevent persistent links.
@@ -417,43 +415,43 @@ pgsql.ignore_notice = 0
 ; http://php.net/pgsql.log-notice
 pgsql.log_notice = 0
 EOF
-    cat >/etc/php/conf.d/pdo_pgsql.ini <<EOF
+    fi
+    if apk info -q -e php-pdo_pgsql; then
+        cat >/etc/php/conf.d/pdo_pgsql.ini <<EOF
 extension=pdo_pgsql.so
 EOF
-else
-    rm -f /etc/php/conf.d/pqsql.ini
-    rm -f /etc/php/conf.d/pdo_pgsql.ini
+    fi
 fi
 
 # -----------------------------------------------------------------------------
 # SQLite3
+rm -f /etc/php/conf.d/sqlite3.ini
+rm -f /etc/php/conf.d/pdo_sqlite.ini
 if [ "$PHP_EXT_SQLITE3" = "yes" ] ; then
-	if ! apk info -q -e php-sqlite3; then
-		apk add -q php-sqlite3
-	fi
-	if ! apk info -q -e php-pdo_sqlite; then
-		apk add -q php-pdo_sqlite
-	fi
-    cat >/etc/php/conf.d/sqlite3.ini <<EOF
+    apk info -q -e php-sqlite3 || apk add -q php-sqlite3
+    apk info -q -e php-pdo_sqlite || apk add -q php-pdo_sqlite
+	if apk info -q -e php-sqlite3; then
+        cat >/etc/php/conf.d/sqlite3.ini <<EOF
 extension=pdo_sqlite.so
 [sqlite3]
 ;sqlite3.extension_dir =
 EOF
-    cat >/etc/php/conf.d/pdo_sqlite.ini <<EOF
+    fi
+	if apk info -q -e php-pdo_sqlite; then
+        cat >/etc/php/conf.d/pdo_sqlite.ini <<EOF
 extension=sqlite3.so
 EOF
-else
-    rm -f /etc/php/conf.d/sqlite3.ini
-    rm -f /etc/php/conf.d/pdo_sqlite.ini
+    fi
+
 fi
 
 # -----------------------------------------------------------------------------
 # SOAP
+rm -f /etc/php/conf.d/soap.ini
 if [ "$PHP_EXT_SOAP" = "yes" ] ; then
-	if ! apk info -q -e php-soap; then
-		apk add -q php-soap
-	fi
-	cat >/etc/php/conf.d/soap.ini <<EOF
+    apk info -q -e php-soap || apk add -q php-soap
+	if apk info -q -e php-soap; then
+	    cat >/etc/php/conf.d/soap.ini <<EOF
 extension=soap.so
 [soap]
 ; Enables or disables WSDL caching feature.
@@ -469,37 +467,34 @@ soap.wsdl_cache_ttl=86400
 ; Sets the size of the cache limit. (Max. number of WSDL files to cache)
 soap.wsdl_cache_limit = 5
 EOF
-else
-    rm -f /etc/php/conf.d/soap.ini
+    fi
 fi
 
 # -----------------------------------------------------------------------------
 # GD
+rm -f /etc/php/conf.d/gd.ini
 if [ "$PHP_EXT_GD" = "yes" ] ; then
-	if ! apk info -q -e php-gd; then
-		apk add -q php-gd
-	fi
-	cat >/etc/php/conf.d/gd.ini <<EOF
+    apk info -q -e php-gd || apk add -q php-gd
+	if apk info -q -e php-gd; then
+	    cat >/etc/php/conf.d/gd.ini <<EOF
 extension=gd.so
 EOF
-else
-    rm -f /etc/php/conf.d/gd.ini
+    fi
 fi
 
 # -----------------------------------------------------------------------------
 # LDAP
+rm -f /etc/php/conf.d/ldap.ini
 if [ "$PHP_EXT_LDAP" = "yes" ] ; then
-	if ! apk info -q -e php-ldap; then
-		apk add -q php-ldap
-	fi
-    cat >/etc/php/conf.d/ldap.ini <<EOF
+    apk info -q -e php-ldap || apk add -q php-ldap
+	if apk info -q -e php-ldap; then
+        cat >/etc/php/conf.d/ldap.ini <<EOF
 extension=ldap.so
 [ldap]
 ; Sets the maximum number of open links or -1 for unlimited.
 ldap.max_links = -1
 EOF
-else
-    rm -f /etc/php/conf.d/ldap.ini
+    fi
 fi
 
 # -----------------------------------------------------------------------------
@@ -509,10 +504,9 @@ rm -f /etc/php/conf.d/xcache.ini
 rm -f /etc/php/conf.d/memcache.ini
 
 if [ "$PHP_EXT_CACHE" = "apc" ] ; then
-	if ! apk info -q -e php-apc; then
-		apk add -q php-apc
-	fi
-    cat >/etc/php/conf.d/apc.ini <<EOF
+    apk info -q -e php-apc || apk add -q php-apc
+	if apk info -q -e php-apc; then
+        cat >/etc/php/conf.d/apc.ini <<EOF
 extension=apc.so
 apc.enabled=1
 ;apc.shm_segments=1
@@ -523,19 +517,18 @@ apc.enabled=1
 apc.mmap_file_mask=/tmp/apc.XXXXXX
 ;apc.enable_cli=1
 EOF
+    fi
 elif [ "${PHP_EXT_CACHE}" = "xcache" ] ; then
-#  later available:
-#	if ! apk info -q -e php-xcache; then
-#		apk add -q php-xcache
-#	fi
-    cat >/etc/php/conf.d/xcache.ini <<EOF
-
+    apk info -q -e php-xcache || apk add -q php-xcache
+	if apk info -q -e php-xcache; then
+        cat >/etc/php/conf.d/xcache.ini <<EOF
+extension=xcache.so
 EOF
+    fi
 elif [ "${PHP_EXT_CACHE}" = "memcache" ] ; then
-	if ! apk info -q -e php-memcache; then
-		apk add -q php-memcache
-	fi
-    cat >/etc/php/conf.d/memcache.ini <<EOF
+    apk info -q -e php-memcache || apk add -q php-memcache
+	if apk info -q -e php-memcache; then
+        cat >/etc/php/conf.d/memcache.ini <<EOF
 extension=memcache.so
 ;memcache.allow_failover="1"
 ;memcache.max_failover_attempts="20"
@@ -551,11 +544,20 @@ extension=memcache.so
 ;memcache.compress_threshold=20000
 ;memcache.lock_timeout=15
 EOF
-
-
+    fi
 fi
 
-
+# -----------------------------------------------------------------------------
+# JSON
+rm -f /etc/php/conf.d/json.ini
+if [ "$PHP_EXT_JSON" = "yes" ] ; then
+	apk info -q -e php-json || apk add -q php-json
+	if apk info -q -e php-json; then
+        cat >/etc/php/conf.d/json.ini <<EOF
+extension=json.so
+EOF
+    fi
+fi
 
 # =============================================================================
 # Restart apache
